Improve call handling and add callbacks
In this commit, we added some useful callbacks, such as callback for answered call, callback for callstate change and also callback for received frames aswell as callback for when the call is hanged up with the reason. We expect that those callbacks will improve the usage of the ap. And more of these will be added later.